% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{compute_transdices}
\alias{compute_transdices}
\title{A function to create indices by the standard transformation, aggregation and transformation
process}
\usage{
compute_transdices(vars, dt, std_funs, agg_fun, index_fun)
}
\arguments{
\item{vars}{a character or list, the variable name(s) of the numeric variables}

\item{dt}{\code{data.frame}/\code{data.table}, contains the set of variables}

\item{std_funs}{the transformation function to be applied to each variable.
length(std_fun) should be 1 if the same function is to be applied to all \code{vars}, otherwise,
length(std_fun) must be equal to length(vars)}

\item{agg_fun}{the aggregation function to combine transformed \code{vars}}

\item{index_fun}{the function applied to the combined transformed variables typically used
to convert the result to the original scale of \code{std_funs}.}
}
\description{
This function applies some transformation function either user defined or standard R function
to a set a variables (could be the same function or a list of functions equal to the length of
variables) and then applies an aggregation function of choice to the set of transformed
indicators and then finally applies a transformation to the result of the aggregation
}
