% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packprocess_funs.R
\name{gen_var_documentation}
\alias{gen_var_documentation}
\title{A function to document dataframe variables with their labels}
\usage{
gen_var_documentation(
  dt,
  file_path,
  title = "Mat's Governance Participation Indicators Dataset",
  description =
    "This dataset contains various indicators related to democracy, governance, social participation and economic conditions across multiple countries and years.  The variables come from sources such as V-Dem, Gallup, and QoG, covering aspects like political engagement, corruption, civil liberties, and electoral integrity.",
  format = "Data frame with indicators, unit of observation country-year",
  source = "V-DEM, Gallup & VOG",
  dataname = "cloutier_dt"
)
}
\arguments{
\item{dt}{a data.frame object with labelled variables}

\item{file_path}{a character of length 1, the file path to a ".R" script
to be created for the documentation (preferably in the R folder)}

\item{title}{character, the title of the dataset}

\item{description}{character, the description of the dataset}

\item{format}{character, the character of the dataset}

\item{source}{character, the source information for the data}

\item{dataname}{character, the name of the data.frame object stored within
the data folder now being documented}
}
\description{
This function will taken any data.frame object with labelled variables and
automatically generate the Variables section of the documentation.
}
\details{
See roxygen documentation for full information on the parameters
}
