% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalcomp.R
\docType{data}
\name{globalcomp_dt}
\alias{globalcomp_dt}
\title{Mat's Governance Participation Indicators Dataset (Global Aggregates)}
\format{
Data frame with global aggregate indicators, unit of observation year
}
\source{
V-DEM, Gallup & VOG indicators, global aggregates
\describe{
\item{year}{Year}
\item{social_capital}{Social Capital}
\item{absence_exclusion}{Absence of Exclusion}
\item{absence_capture}{Absence of Elite Capture}
\item{vdeminformal}{Scaled sum of variables v2x_freexp_altinf and v2caassemb}
\item{btiinformal}{Scaled sum of variables bti_aar and bti_foe}
\item{informal_channels}{Informal Channels}
\item{vdeminstitutional}{Scaled sum of variables v2xel_frefair and v2xel_locelec}
\item{btiinstitutional}{Scaled sum of variables bti_ffe and ibp_cat}
\item{institutional_channels}{Institutional Channels}
\item{vdemcso}{Scaled sum of variables v2csprtcpt and v2cscnsult}
\item{bticso}{Scaled sum of the variables bti_ig and bti_csp}
\item{intermediary_channels}{Intermediary Channels}
\item{civil_capacity}{Civil Capacity}
\item{quality_interface}{Quality of the Citizen-State Interface}
\item{resilience}{Resilience of the Social Contract}
\item{v2cagenmob}{To what extent are elections free and fair?}
\item{percept_civicspace}{Perceptions of the Civic Space}
\item{natgov_confidence}{Confidence in National Government}
}
}
\usage{
globalcomp_dt
}
\description{
This data computes the global aggregates from the Cloutier Dataset
}
\examples{
globalcomp_dt
}
\keyword{datasets}
