% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcomp.R
\docType{data}
\name{regcomp_dt}
\alias{regcomp_dt}
\title{Mat's Governance Participation Indicators Dataset (Regional Aggregates)}
\format{
Data frame with regional aggregate indicators, unit of observation region-year
}
\source{
V-DEM, Gallup & VOG indicators, regional aggregates
\describe{
\item{year}{Social Capital}
\item{wbregion}{Absence of Exclusion}
\item{social_capital}{Absence of Elite Capture}
\item{absence_exclusion}{Scaled sum of variables v2x_freexp_altinf and v2caassemb}
\item{absence_capture}{Scaled sum of variables bti_aar and bti_foe}
\item{vdeminformal}{Informal Channels}
\item{btiinformal}{Scaled sum of variables v2xel_frefair and v2xel_locelec}
\item{informal_channels}{Scaled sum of variables bti_ffe and ibp_cat}
\item{vdeminstitutional}{Institutional Channels}
\item{btiinstitutional}{Scaled sum of variables v2csprtcpt and v2cscnsult}
\item{institutional_channels}{Scaled sum of the variables bti_ig and bti_csp}
\item{vdemcso}{Intermediary Channels}
\item{bticso}{Civil Capacity}
\item{intermediary_channels}{Quality of the Citizen-State Interface}
\item{civil_capacity}{Resilience of the Social Contract}
\item{quality_interface}{To what extent are elections free and fair?}
\item{resilience}{Perceptions of the Civic Space}
\item{v2cagenmob}{Confidence in National Government}
\item{percept_civicspace}{Social Capital}
\item{natgov_confidence}{Absence of Exclusion}
}
}
\usage{
regcomp_dt
}
\description{
This data computes the regional aggregates from the Cloutier Dataset
}
\examples{
regcomp_dt
}
\keyword{datasets}
